update pom and working

+++ src/main/java/com/ProjectManagement/ProjectManagement/repository/TestRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TestRepo extends CrudRepository<TestEntity, Integer> {
+public interface TestRepo extends CrudRepository<TestEntity, Long> {
 }
